requirements.txt
fastapi[standard]==0.115.11
pdf2image==1.17.0
pytesseract==0.3.13
opencv-python-headless==4.11.0.86
python-dotenv==1.0.1
huggingface_hub==0.29.3